requirements.txt
chess
keras
numpy
pyqt5_sip
tensorflow
